% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_complexcomp_package.R
\name{play_complexcomp}
\alias{play_complexcomp}
\title{play_complexcomp}
\usage{
play_complexcomp(
  evolution,
  lifespan,
  num_scientists,
  max_scientists_per_q,
  startup_cost,
  sample_cost,
  exp_shape,
  decay,
  b_neg,
  abandon_prob,
  ss,
  min_sample_size,
  max_sample_size
)
}
\arguments{
\item{evolution}{A boolean. When set to TRUE, sample sizes are fixed.}

\item{lifespan}{lifespan; smaller lifespan = more noise}

\item{num_scientists}{num_scientists}

\item{max_scientists_per_q}{max_scientists_per_q}

\item{startup_cost}{startup_cost}

\item{sample_cost}{sample_cost}

\item{exp_shape}{exp_shape}

\item{decay}{decay}

\item{b_neg}{b_neg}

\item{abandon_prob}{abandon_prob}

\item{ss}{fixed sample size}

\item{min_sample_size}{min_sample_size for when sample
sizes are drawn from distribution}

\item{max_sample_size}{max_sample_size for when sample
sizes are drawn from distribution}
}
\value{

}
\description{
Function to play scientists against one another,
including measure of number of scoops
}
