% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_complexcomp_package.R
\name{play_complexcomp}
\alias{play_complexcomp}
\title{play_complexcomp}
\usage{
play_complexcomp(
  evolution,
  lifespan,
  ss,
  max_players_per_q,
  startup_cost,
  sample_cost,
  exp_shape,
  decay,
  b_neg,
  abandon_prob
)
}
\arguments{
\item{evolution}{evolution}

\item{lifespan}{lifespan}

\item{ss}{ss}

\item{max_players_per_q}{max_players_per_q}

\item{startup_cost}{startup_cost}

\item{sample_cost}{sample_cost}

\item{exp_shape}{exp_shape}

\item{decay}{decay}

\item{b_neg}{b_neg}

\item{abandon_prob}{abandon_prob}
}
\value{

}
\description{
Function to play scientists against one another,
including measure of number of scoops
}
